moved exp_queue to utils
