+++ bugbug/repository.py
@@ -14,7 +14,6 @@ import os
 import pickle
 import sys
 import threading
-from collections import deque
 from datetime import datetime
 
 import hglib
@@ -345,59 +344,6 @@ def get_revs(hg, rev_start=0, rev_end="tip"):
     return x.splitlines()
 
 
-class exp_queue:
-    def __init__(self, start_day, maxlen, default):
-        self.list = deque([default] * maxlen, maxlen=maxlen)
-        self.start_day = start_day - (maxlen - 1)
-        self.default = default
-
-    def __deepcopy__(self, memo):
-        result = exp_queue.__new__(exp_queue)
-
-        # We don't need to deepcopy the list, as elements in the list are immutable.
-        result.list = self.list.copy()
-        result.start_day = self.start_day
-        result.default = self.default
-
-        return result
-
-    @property
-    def last_day(self):
-        return self.start_day + (self.list.maxlen - 1)
-
-    def __getitem__(self, day):
-        assert (
-            day >= self.start_day
-        ), f"Can't get a day ({day}) from earlier than start day ({self.start_day})"
-
-        if day < 0:
-            return self.default
-
-        if day > self.last_day:
-            return self.list[-1]
-
-        return self.list[day - self.start_day]
-
-    def __setitem__(self, day, value):
-        if day == self.last_day:
-            self.list[day - self.start_day] = value
-        elif day > self.last_day:
-            last_val = self.list[-1]
-            # We need to extend the list except for 2 elements (the last, which
-            # is going to be the same, and the one we are adding now).
-            range_end = min(day - self.last_day, self.list.maxlen) - 2
-            if range_end > 0:
-                self.list.extend(last_val for _ in range(range_end))
-
-            self.start_day = day - (self.list.maxlen - 1)
-
-            self.list.append(value)
-        else:
-            assert False, "Can't insert in the past"
-
-        assert day == self.last_day
-
-
 def calculate_experiences(commits, first_pushdate, save=True):
     print(f"Analyzing experiences from {len(commits)} commits...")
 
@@ -429,7 +375,7 @@ def calculate_experiences(commits, first_pushdate, save=True):
             experiences[exp_type][commit_type] = {}
 
         if item not in experiences[exp_type][commit_type]:
-            experiences[exp_type][commit_type][item] = exp_queue(
+            experiences[exp_type][commit_type][item] = utils.ExpQueue(
                 day, EXPERIENCE_TIMESPAN + 1, default
             )
